+++ server/src/GameLogic/Weapon.cs
@@ -2,8 +2,8 @@ namespace Thuai.Server.GameLogic;
 
 public class Weapon
 {
-    public double AttackSpeed { get; set; } = Constants.INITIAL_ATTACK_SPEED;
-    public double BulletSpeed { get; set; } = Constants.INITIAL_BULLET_SPEED;
+    public float AttackSpeed { get; set; } = Constants.INITIAL_ATTACK_SPEED;
+    public float BulletSpeed { get; set; } = Constants.INITIAL_BULLET_SPEED;
     public bool IsLaser { get; set; } = false;
     public bool AntiArmor { get; set; } = false;
     public int Damage { get; set; } = Constants.INITIAL_DAMAGE;
